Moved AbstractSyntaxTree into file
Removed AbstractSyntaxTree from the Language module file and placed into it's own source file. The Language module now contains require statements to Rule, RuleSeq and AbstractSyntaxTree as well as the module declaration.
